server/requirements.txt
asgiref==3.8.1
boto3==1.35.29
botocore==1.35.29
cffi==1.17.1
cryptography==43.0.1
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jmespath==1.0.1
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.20